refactor: replace mock.module test-isolation workaround with spyOn
Bun's mock.module() is process-global and retroactive: once a test file
mocks a first-party module path, that mock replaces the module for every
test file in the same process and is not undone by mock.restore(). To get
the real implementations, the suite had split production code into
auth-poll.ts and credential-store-impl.ts, with auth.ts/credential-store.ts
reduced to thin delegating wrappers, purely so the "real" tests could import
an un-mocked path. The isolation depended on test-file execution order,
which is the actual source of the flakiness.

Fix the root cause instead of contorting the module layout:

- Merge auth-poll.ts back into auth.ts and credential-store-impl.ts back
  into credential-store.ts; delete the wrapper files.
- Convert the three leaky mockers (login-flow, init-project, plugin/install)
  from mock.module() to `import * as mod` + spyOn in beforeEach, restored via
  mock.restore() in afterEach. spyOn is per-test and auto-restored, and it
  reaches both static named-import and dynamic `await import()` consumers
  because it mutates the single shared module instance (same pattern already
  used for pathsMod.findProjectRoot). Third-party mocks (inquirer,
  node:readline) keep using mock.module.
- Consolidate tests: fold auth-poll.test.ts into auth.test.ts; drop the stub
  credential-store re-export test and move the real impl tests into
  credential-store.test.ts.
- Document the rule in ARCH-005 (DO/DON'T, good/bad example, compliance
  item) so the workaround does not reappear.
