+++ src/helpers/auth.ts
@@ -8,23 +8,8 @@
  * credential helpers (macOS Keychain, Windows Credential Manager, libsecret).
  */
 
-// Re-export pollForAccessToken as a wrapper (NOT a live re-export) so that
-// mock.module("auth") in login-flow.test.ts does NOT follow the binding chain
-// into auth-poll.ts. A live `export { X } from "./Y"` creates a binding that
-// Bun's mock.module replaces at the source, poisoning auth-poll.ts for other
-// test files. A wrapper function is its own binding — mocking auth.ts replaces
-// the wrapper, leaving auth-poll.ts's binding untouched.
-import { pollForAccessToken as pollForAccessTokenImpl } from "./auth-poll";
 import { logDebug } from "./log";
 import { SignupRequiredError, isSignupRequiredError } from "./signup";
-
-export async function pollForAccessToken(
-  deviceCode: string,
-  interval: number,
-  expiresIn: number
-): Promise<string> {
-  return await pollForAccessTokenImpl(deviceCode, interval, expiresIn);
-}
 
 // ---------------------------------------------------------------------------
 // Constants
@@ -32,6 +17,7 @@ export async function pollForAccessToken(
 
 const PLUGINS_API = "https://plugins.archgate.dev";
 const GITHUB_DEVICE_CODE_URL = "https://github.com/login/device/code";
+const GITHUB_DEVICE_TOKEN_URL = "https://github.com/login/oauth/access_token";
 
 /**
  * GitHub OAuth App client ID for the archgate CLI (public client — no secret).
@@ -50,6 +36,27 @@ interface DeviceCodeResponse {
   expires_in: number;
   interval: number;
 }
+
+interface DeviceTokenSuccessResponse {
+  access_token: string;
+  token_type: string;
+  scope: string;
+}
+
+interface DeviceTokenPendingResponse {
+  error: "authorization_pending" | "slow_down";
+  error_description?: string;
+}
+
+interface DeviceTokenErrorResponse {
+  error: "expired_token" | "access_denied" | string;
+  error_description?: string;
+}
+
+type DeviceTokenResponse =
+  | DeviceTokenSuccessResponse
+  | DeviceTokenPendingResponse
+  | DeviceTokenErrorResponse;
 
 // ---------------------------------------------------------------------------
 // GitHub Device Flow
@@ -78,6 +85,72 @@ export async function requestDeviceCode(): Promise<DeviceCodeResponse> {
   const data = (await response.json()) as DeviceCodeResponse;
   logDebug("Device code received, expires in:", data.expires_in, "seconds");
   return data;
+}
+
+/**
+ * Step 2: Poll GitHub until the user authorizes (or the code expires).
+ * Returns the GitHub access token on success.
+ */
+export async function pollForAccessToken(
+  deviceCode: string,
+  interval: number,
+  expiresIn: number
+): Promise<string> {
+  const deadline = Date.now() + expiresIn * 1000;
+  let pollInterval = interval;
+  logDebug(
+    "Starting device code polling, deadline:",
+    new Date(deadline).toISOString()
+  );
+
+  /* oxlint-disable no-await-in-loop -- sequential polling is required by RFC 8628 */
+  while (Date.now() < deadline) {
+    await Bun.sleep(pollInterval * 1000);
+
+    const response = await fetch(GITHUB_DEVICE_TOKEN_URL, {
+      method: "POST",
+      headers: {
+        Accept: "application/json",
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({
+        client_id: GITHUB_CLIENT_ID,
+        device_code: deviceCode,
+        grant_type: "urn:ietf:params:oauth:grant-type:device_code",
+      }),
+      signal: AbortSignal.timeout(15_000),
+    });
+
+    if (!response.ok) {
+      throw new Error(`GitHub token poll failed (HTTP ${response.status})`);
+    }
+
+    const data = (await response.json()) as DeviceTokenResponse;
+
+    if ("access_token" in data) {
+      logDebug("Access token received");
+      return data.access_token;
+    }
+
+    if ("error" in data) {
+      if (data.error === "authorization_pending") {
+        logDebug("Authorization pending, retrying in", pollInterval, "seconds");
+        continue;
+      }
+      if (data.error === "slow_down") {
+        pollInterval += 5;
+        logDebug("Slow down requested, new interval:", pollInterval, "seconds");
+        continue;
+      }
+      // expired_token, access_denied, or other terminal error
+      throw new Error(
+        data.error_description ?? `GitHub authorization failed: ${data.error}`
+      );
+    }
+  }
+  /* oxlint-enable no-await-in-loop */
+
+  throw new Error("Device code expired. Please try again.");
 }
 
 interface GitHubUserInfo {
